PR-09: extract backends/ and backends/wayland/
Highest-macro-risk PR — wayland_scanner generators relocated into a
nested module. PR-00 confirmed CARGO_MANIFEST_DIR-relative path
resolution; the literal "src/protocols/..." path strings ride
along unchanged.

- backends/mod.rs: RawFdWatcher (shared by wayland and x11 per D05);
  pub(crate) mod wayland; pub(crate) use wayland::*.
- backends/wayland/mod.rs: ToplevelWindow, WaylandState + impl,
  WaylandProtocol, run_wayland and the four wayland helpers
  (resolve_wayland_socket_path, connect_wayland_with_display_override,
  query_wayland_active_window, wayland_query_count). Submodule
  re-exports per D24.
- backends/wayland/protocols.rs: mod cosmic_workspace + mod cosmic_toplevel
  with wayland_scanner generators. Intra-module crate::cosmic_workspace
  paths rewritten to super::/super::super:: per D03.
- backends/wayland/dispatch_{common,wlr,cosmic}.rs: split-out Dispatch
  impls (8 total) for cleaner per-protocol diffs.

main.rs: 4026 -> 3526 LOC. supervisor/mod.rs's run_wayland call site
updated to crate::backends::wayland::run_wayland.

Verification: cargo build clean (warnings 8 -> 18, all unused-import
drift; cleanup deferred); cargo test --bin kanata-switcher 261/0.
