+++ src/daemon/backends/wayland/dispatch_common.rs
@@ -1,0 +1,31 @@
+use super::WaylandState;
+use wayland_client::{
+    Connection as WaylandConnection, Dispatch, QueueHandle,
+    globals::GlobalListContents,
+    protocol::wl_registry,
+};
+
+impl Dispatch<wl_registry::WlRegistry, GlobalListContents> for WaylandState {
+    fn event(
+        _: &mut Self,
+        _: &wl_registry::WlRegistry,
+        _: wl_registry::Event,
+        _: &GlobalListContents,
+        _: &WaylandConnection,
+        _: &QueueHandle<Self>,
+    ) {
+    }
+}
+
+// Dispatch for wl_output (referenced by toplevel protocol)
+impl Dispatch<wayland_client::protocol::wl_output::WlOutput, ()> for WaylandState {
+    fn event(
+        _: &mut Self,
+        _: &wayland_client::protocol::wl_output::WlOutput,
+        _: wayland_client::protocol::wl_output::Event,
+        _: &(),
+        _: &WaylandConnection,
+        _: &QueueHandle<Self>,
+    ) {
+    }
+}
