+++ src/daemon/backends/wayland/dispatch_wlr.rs
@@ -1,0 +1,69 @@
+use super::{ToplevelWindow, WaylandState};
+use wayland_client::{Connection as WaylandConnection, Dispatch, Proxy, QueueHandle};
+use wayland_protocols_wlr::foreign_toplevel::v1::client::{
+    zwlr_foreign_toplevel_handle_v1::{self, ZwlrForeignToplevelHandleV1},
+    zwlr_foreign_toplevel_manager_v1::{self, ZwlrForeignToplevelManagerV1},
+};
+
+impl Dispatch<ZwlrForeignToplevelManagerV1, ()> for WaylandState {
+    fn event(
+        state: &mut Self,
+        _: &ZwlrForeignToplevelManagerV1,
+        event: zwlr_foreign_toplevel_manager_v1::Event,
+        _: &(),
+        _: &WaylandConnection,
+        _: &QueueHandle<Self>,
+    ) {
+        if let zwlr_foreign_toplevel_manager_v1::Event::Toplevel { toplevel } = event {
+            state
+                .windows
+                .insert(toplevel.id(), ToplevelWindow::default());
+        }
+    }
+
+    wayland_client::event_created_child!(WaylandState, ZwlrForeignToplevelManagerV1, [
+        zwlr_foreign_toplevel_manager_v1::EVT_TOPLEVEL_OPCODE => (ZwlrForeignToplevelHandleV1, ())
+    ]);
+}
+
+impl Dispatch<ZwlrForeignToplevelHandleV1, ()> for WaylandState {
+    fn event(
+        state: &mut Self,
+        handle: &ZwlrForeignToplevelHandleV1,
+        event: zwlr_foreign_toplevel_handle_v1::Event,
+        _: &(),
+        _: &WaylandConnection,
+        _: &QueueHandle<Self>,
+    ) {
+        match event {
+            zwlr_foreign_toplevel_handle_v1::Event::AppId { app_id } => {
+                if let Some(w) = state.windows.get_mut(&handle.id()) {
+                    w.app_id = app_id;
+                }
+            }
+            zwlr_foreign_toplevel_handle_v1::Event::Title { title } => {
+                if let Some(w) = state.windows.get_mut(&handle.id()) {
+                    w.title = title;
+                }
+            }
+            zwlr_foreign_toplevel_handle_v1::Event::State {
+                state: handle_state,
+            } => {
+                let activated = zwlr_foreign_toplevel_handle_v1::State::Activated as u8;
+                if handle_state.contains(&activated) {
+                    state.active_window = Some(handle.id());
+                } else if state.active_window.as_ref() == Some(&handle.id()) {
+                    // Window lost activation - clear active_window
+                    state.active_window = None;
+                }
+            }
+            zwlr_foreign_toplevel_handle_v1::Event::Closed => {
+                state.windows.remove(&handle.id());
+                if state.active_window.as_ref() == Some(&handle.id()) {
+                    state.active_window = None;
+                }
+            }
+            _ => {}
+        }
+    }
+}
